Desktop SX: Fix target section title initialization and update
Until now, when the target title is updated in desktop article
translations, an autosave request is triggered. However, this is not
the case for desktop section translations (even if it should be).

Additionally, when the target section title is changed, and user tries
to leave the page a message about unsaved changes that will be lost,
is displayed to the user. This happens even for cases where an autosave
request has been sent and section title along with translated contents
have been properly stored in the database.

To fix above issues, this patch properly initializes and updates the
saved target title for section translations. The patch also implements
the target section title update handler to send autosave requests when
the target section title is updated.

Change-Id: I5d68da58c965981e2f9a83f5dfbdeeb9255dcc2d