CX Special Page: Use "requireNamedUser" instead of "requireLogin"
Currently, for requests to Special:ContentTranslation page, when
the URL "campaign" parameter is not valid, we want to redirect
the user to the login page, if the user is anonymous.

Since the (newly introduced) temp users should be treated as anon.
users for now, we should redirect such users to the login page for
the above scenario, too.

This patch replaces the call of "requireLogin" method of the
"SpecialPage" class, with the "requireNamedUser" method of the same
class, to properly handle temp users, too.

Bug: T344771
Change-Id: I8d9bb5515c590df19bf19a6e38823e8b1ee9604d