SX UI lib: Add "persistent" prop to MWDialog SFC
Currently, when a MwDialog instance is open, and the user clicks on
the overlay of the dialog, the dialog always closes. However, this
may not be the desired for some cases.

For this reason, this patch adds the "persistent" boolean prop. When
this prop is set to false (which is the default), the dialog will be
closed on overlay click. However, if set to true, the dialog will not
close when clicking on the overlay.

Change-Id: I0613b74c34290b928d398c25a9c18fe19383d1a7