CX contenttranslation list: Move translation unit fetching to method
When a query request to the "contenttranslation" list is send with
the "translationid" parameter set, the API returns a response
containing the translation data, as well as the translation corpora
units.

This patch moves the code that fetching the translation corpora units
to a separate method, named "addUnitsAndCategoriesToTranslation", for
better code organisation and code reusability.

Change-Id: I7337ead534211bcc03fc0c3bd3b145c2d1fd9bdc