CX contenttranslation list: Add 'desktop-editor-draft' usecase
This patch adds the "desktop-editor-draft" usecase for the
'contenttranslation' list of the 'query' API action.

The "serveDesktopEditorDraft" method that serves this usecase is
a replacement of the old "find" method, that fetched and returned
the draft translation or a conflicting translation. The difference
is that "serveDesktopEditorDraft" also fetches the translation corpora
units of the draft translation (if such translation is found).

This way we combine the two different requests that were sent to fetch
the draft translation data (information + translation units), into one.

Bug: T348629
Change-Id: I868dbd4fabdcf4547c9e5eef83fef2520d7d9f76