CX: Move Translation::save method to TranslationStore::saveTranslation
This patch moves the "save" method of the Translation class to the
TranslationStore and renames it to "saveTranslation". The purpose of
this change is to move database access functionality from the
Translation model to the corresponding store, for better separation of
concerns and proper dependency injection.

The "save" method used to call the "create" and "update" methods inside
Translation class, so these methods are also moved and renamed to
"insertTranslation" and "updateTranslation", respectively.

This patch also updates all occurrences of the above methods and renames
the "lastSaveWasCreate" property of Translation class to "isNew", so
that the property name is more self-explanatory.

Change-Id: Iae7fdd412b1c9f4346896d27aa5aa0dc86b5225f