CX: Remove "$user" parameter from TranslationStore::updateTranslation
Until now, TranslationStore::updateTranslation requires the "$user"
parameter, so that it can update the "translation_last_update_by" field
to the user's global id. However, since we allow users to edit only
their own translations, there is no case that this field should be
updated.

In fact, the "translation_last_update_by" is a redundant field that
could/should be removed from "cx_translations" table, as it is always
equal to the "translation_started_by" field. Currently, we use it to
fetch the translator id of the translation, but we could also use the
"translation_started_by" field instead.

This patch also completes a 6-years old TODO item, that is it removes
the update of the "translation_started_by" field, which also didn't add
any value and was introduced as a temporary solution back in 2017, in
this patch: I4ecf067ed75a4b0b7f26e6f00c5d6e23ae54e503.

Finally, since the user is no longer needed in the "updateTranslation"
method, this patch removes it from its parameters.

Change-Id: Ib83b17301dcb3d166dfba8f7a2ccd2340cf5e9b6