CX: Remove BC compatibility to 5-year-old schema change
The "translation_cx_version" field was introduced about 5 years ago,
Back then, as a temporary measure, in order to provide backward
compatibility a check for this field was added, during DB inserts and
updates.

However, this check is no longer relevant as we do not expect any
CX installation to not support this field until now.

Patch where the check was added:
I5e23c8045949a644d073587668e55109153ce43c

Change-Id: Id80ef2fdc148c1f46da7d26818aef1ac83136aee