Replace WikimediaUI Base with Codex design tokens
This patch replaces the WikimediaUI Base library with the new
Codex design tokens library.
The new library is mostly a drop-in replacement for the old one.
But there are deprecated variable names (token keys) that need to be
adapted to the new ones.

Also for code hygiene, we move the prefers reduced motion style rules
into new separate file 'common.less' and make tokens only imported as
Less `(reference)`.

Replacing also a number of static values mainly on properties, where
most other occurrences had already majority of variable usage. See
`font-weight` or `border-radius` for example.
Please note that the only bigger “step back” (to static values) is the
replacement of `color-base--subtle` by `#72777d` as the new
`color-subtle` was agreed on by systems designers to carry darker
`#54595d` value. For now approaching it this way makes most sense for
direct replacement testing, we can work in a separate patch with
Language design if the darker `color-subtle` is desired.

Bug: T331403
Bug: T332120
Bug: T350048
Change-Id: I5a8881fd36d2d4b10ced1836185f15477221481c