SX instrumentation: Fix translationSessionPosition setting
Until now the translation session position inside instrumentation
events, was always set to 1, although it should be auto-incremented.
This happens because the stored session position inside local storage,
is cleared before fetched, and thus the counter is always reset to 0.
To fix this issue, this patch fetches the session position before
clearing the local storage, preserving this way the previously stored
value.

Additionally, the usage of "localStorage" global variable was replaced
with the usage of "mw.storage".

Tests added too.

Change-Id: Iebc1f09881ce2e175414ac0d155eedabfff8bba0