CX Suggestion List: Use the language filter codes for translation start
Some wikis have a different language code and domain code mapping (e.g.
the language code for Norwegian Bokmål is "nb", while the domain code
for the corresponding wiki is "no").

This causes confusion when starting new translations from the suggestion
list, as the suggestions are often fetched from the recommendation tool
which uses the wiki domain codes. However, the language codes of the
fetched suggestions can be different that the selected language code
in the language filter, and thus new translations can start with a
different language code than the selected one.

To fix this issue, this patch uses the language codes from the language
filter, when opening a new translation from the suggestion list.
Additionally, it changes the language pair displayed at the bottom of
the suggestion card, to display the language pair that is selected
in the language filter, as this is the pair that will be used, if the
suggestion is clicked.

Bug: T339091
Change-Id: I819a0e8c23366fe5b3e6847a32c22415e45028ac