CX draft restoration: Fix conflicting translation case
Following the latest changes in this patch:
I18042d48fec987f335ab01d8beaf7d3ff61715aa, when a conflicting translation
exists, instead of a dialog letting the user know about the conflicting
translation by another user, a "Unable to fetch draft translation"
message is being displayed to the user.

This happens because apparently boolean values cannot be added to the
request result, and thus the "hasConflicts" property in the response
payload is empty even for cases when conflicting translation exists and
the property is added to the payload.

This patch fixes this issue by setting the value of the "hasConflicts"
property to "true", as a string. Additionally, here we set the default
argument of the "mw.cx.dm.DraftTranslationDTO" constructor to an empty
object, to avoid errors being thrown when the "translation" property
of the response payload is undefined (which is the case when conflicting
translation exists).

Bug: T348558
Change-Id: Ibd3d47ba09b4e8c6c7c6dbaa6d22d9d4f30adcc0