SX section position calculation: Use "index" property from parse response
When a section translation is being published, the "cxpublishsection"
API uses the SectionPositionCalculator service to calculate the proper
position, where the new section should be positioned. This service
sends a request to the "parse" Action API to fetch the section titles
and positions of the target article (if such article exists), and finds
the position of the first appendix title within these sections.

Until now, the property that was used to get the position of each section
inside the target article was the "number" property of each section
object. However, this property holds a text representation of the
section position, and not an integer, like we would expect. That leads
to wrong section position calculations, especially in wikis/languages
where number are represented with different symbols (see the
corresponding parse request for "gu" wiki below). Even worse, in such
wikis the section position can be calculated to "0", for non-lead
sections, which leads to unexpected behaviour, as we strongly depend
on the assumption that section position equal to "0" means lead section
translation.

This patch replaces the usage of the "number" property with the "index"
property of the section objects, which presents the desired behaviour,
and always holds a valid integer value.

Example parse requests for target section fetching:
1. https://en.wikipedia.org/w/api.php?action=parse&format=json&prop=sections&formatversion=2&page=Football
2. https://gu.wikipedia.org/w/api.php?action=parse&format=json&prop=sections&formatversion=2&page=%E0%AA%AF%E0%AA%BE%E0%AA%A4%E0%AB%8D%E0%AA%B0%E0%AA%BE
Change-Id: I70e4f8750b32878e976f47c33e4efa4217b854f4