CX relevant languages entrypoint: Return promise from fetchSiteMatrix
Currently, although the "fetchSiteMatrix" is supposed to always return
a promise, it only returns a regular array when the "sitematrix"
request has already been sent once. That leads to a javascript error:
"TypeError: fetchSiteMatrix(...).then is not a function", as the code
expects a promise to call "then" on it, but it gets an array instead.
This happens in cases when the entrypoint is not rendered the first
time the code executes, i.e. when no relevant language is missing.

This patch fixes this issue by return a trivial promise that resolves
to the "wikiLanguageCodes" array, for cases when the "sitematrix"
request has already been sent.

Bug: T349931
Change-Id: I9bb81cb312f11bafbcb99300ee05f20cbf48f623