SX: Avoid to store "mt" translation units when no MT provider is given
Currently, the "parallelCorporaMTContent" getter inside subSection model
returns a non-empty HTML string for block templates even when no MT
provider exists, which is the case of a restored translation that used
"source/original" or "empty" MT providers.

This leads to validation errors, as the origin of a translation unit
that is going to be saved to the "cx_corpora" table cannot be empty.

This patch updates the "parallelCorporaMTContent" getter to return
null when the "mtProviderUsed" property of a section is empty.

Screenshot of the validation error in console:
https://phabricator.wikimedia.org/F41534538

Change-Id: I0d477d4c7c57b51039c28dfd8434bbbdecb50e8f