CX translation target URL creation: Use target title from the request
Until now, the creation of the target URL of a translation, involved
the SiteMapper::getTargetTitle method. This method has been outdated
for a long time now, but it kept working by accident until PHP 8.

Until PHP 7, this method is only returning the given target title, when
"ContentTranslationTranslateInTarget" configuration parameter is set
to "true". In production, the "ContentTranslationTargetNamespace"
configuration parameter is set to 0, and the comparison 'Main' == 0,
yielded "true" because "switch" does loose comparison in PHP 7.

That changed in PHP 8, and the method returns the target title
prefixed with the namespace number, resulting into weird target URLs
(e.g. "0:My target title").

To fix the above issues, this patch entirely removes the obsolete
"getTargetTitle" method, and uses the target title as given in the
publish request payload, as this was the behaviour until now. The
creation of the target URLs will be improved in following patches.

Change-Id: I730db07216cb1273df30942da7ff1903c192a96c