CX: Use the constructed Title object for target URL creation
Currently, the raw target title as given in the publish request
payload, is used for the creation of the translation target URL.
However, in order for the URL to be valid, we need to pass the
same page title that is also used for the page creation/edit.

This patch updates the "cxpublish" and "cxpublishsection" APIs
to properly create the target URLs, using the "getPrefixedDBKey"
method of the targetTitle object.

Change-Id: I4c1a8d90d612ad47fb9603c5502c6c24705f2144