CX: Fix SandboxTitleMaker to always include username in the title
Until now, the SandboxTitleMaker::makeSandboxTitle didn't include the
username in the created title, when the SandboxLink extension is not
loaded, leading to target titles in the following wrong format:
"User:My target title". This went unnoticed until now because the
SandboxLink extension is loaded in production, but the issue should
still be fixed, even for the development/test environments.

Change-Id: I78839a2d37a248b141d8eb339fee030bd79234b0