CX: Return target URL from "cxpublishsection" API and use it in SX
Until now, the "cxpublishsection" endpoint was returning the target
title of the translation inside the response payload. This target title
was used by the UI, to create the URL, to which the user should be
redirected after successful publishing.

This target title can be different from the title that is given as
request payload to this endpoint, mostly when the publishing happens
to the user sandbox, and this is the reason why the target title was
included in the response.

However, since the returned target title is only used for redirection
to the target title, we can return the target URL instead, and use it
without further manipulations.

This patch replaces the "targettitle" property in the response payload,
with the "targeturl" property, which is what is really needed by the
UI.

Until a new CX build is created, the old "targettitle" is left in the
response payload, so that we don't break the functionality because
of inconsistent SX UI and SX backend.

Change-Id: Ic3b17310a6345a62219dd1449d3f09e31e226821