CX: Support section publishing to sandbox
CX allows publishing in the user namespace by just changing the target
title (which is editable in article translations), to include the user
namespace prefix. However, in section translations the target title
article is not editable in desktop editor, only the target section
title is.

That means we need to follow a different approach to support sandbox
publishing for sections on desktop. The "cxpublishsection" API action
expects an "issandbox" parameter that defines whether the section
should be published in the main or the user namespace.

This patch updates the way we store, update and get the namespace for
publishing and it fixes the "getPublishNamespace" getter to return the
proper value to populate that "issandbox" parameter, when user selects
sandbox publishing.

Bug: T351318
Change-Id: I2515a03f5bedfda19cb61298af479f9239f0f134