SX: Add status filter when searching for a draft translation in store
Currently, we use the "translator/getTranslation" getter to
get a draft translation based on its page title, section title and
language pair. Although, we also check against the "sourceSectionTitle"
property, which only exists in DraftTranslation model, the naming of
the method can be confusing, and the method can even return unexpected
results, if, for example, the given sectionTitle is undefined (that
could also match published translations).

This patch renames the method for clarity and adds a status filter, so
that we explicitly filter only draft translations for this getter.

Change-Id: I6cf8539a06b40e31075e48e8c1f0cd61490f998b