Rename "getSuggestionListLanguagePairUpdater" composable
This patch renames the "getSuggestionListLanguagePairUpdater" method
inside "useLanguageHelper" composable file, to follow the standard
"useX" composable naming style.

Additionally, the method is changed to the standard composable format,
i.e. the new "useSuggestionListLanguagePairUpdate" method should only
be called within "setup" methods/scripts and it will return a function
that can be used to update the language pair for a suggestion list.

Change-Id: I05758bc43988c00b0b59723f14894883c5ecfed8