Rename "getArticleLanguagePairUpdater" composable
This patch renames the "getArticleLanguagePairUpdater" method inside
"useLanguageHelper" composable file, to follow the standard "useX"
composable naming style.

Additionally, the method is changed to the standard composable format,
i.e. the new "useArticleLanguagePairUpdate" method should only be
called within "setup" methods/scripts and it will return a function
that can be used to update the language pair for a selected article,
basically inside the "Confirm a translation" and "Pick a section"
steps.

Change-Id: I97e0f2cbc582746d9370dddd37d56056d90298f4