CX Unified dashboard: Add "useSuggestionSeedsInitialize" composable
The implementation of the "fetching suggestion seeds" functionality was
very complicated until now, involving different seeds for page and
section suggestions, and 3 different vuex actions to support it.

The reason for that, is that we initially thought about supporting
multiple "seed providers", that is multiple ways to fetch seeds for
the suggestions (e.g. all published translations for the given source
wiki, user translations, etc). However, we never really enabled any
other seed provider, except for the published translations in the
source wiki, and we do not have any such plan either.

For the above reasons, this patch simplifies the suggestions seed
fetching functionality, by:
1. using a single state variable for both section and page suggestions,
as there is no reason for using different seeds for this. We are
currently fetching 200 seeds on a single request, which are more
than enough to serve all needed page and section suggestions. Every
time a seed is used (for page/section suggestion fetching), it's being
removed from this vuex state array variable.

Using a single state variable means that we also reduce the number of
seed fetching requests from 2 (one for page and one for section
suggestions) to a single one.
2. Removing all redundant vuex actions and keeping only one to support
this functionality ("getSuggestionSeed").
3. Simplify "suggestionSeedCollection" DTO class, by removing all the
complex "multiple seed providers" functionality.
4. Move the logic to a separate composable that can be more easily
read/expanded/tested.

Change-Id: I42eb2f1cdd8765e209d4f4ff1c9e7c519752c90f