SX: Use v-if instead of v-show to only display published list on demand
Currently, when the unified dashboard loads, requests are being sent to
cxserver to fetch section suggestions for published translations (one
request for each translation). This leads to several non-critical
requests being dispatched in parallel with the critical requests for
the dashboard initialization, and consuming resources. This happens
because we render the published translation list - just with display
"hidden" - even if it's not displayed until the user clicks on the
"published" tab.

This patch uses "v-if" to conditionally display the published list,
instead of "v-show". "v-if" doesn't render the list, until it should
be displayed.

Because "v-if" removes the component from the DOM when its condition
becomes false (for example when clicking on the published tab and
then on the in-progress tab), we need to store the fetched section
suggestions for the published translations, otherwise all the requests
would be re-sent every time the published list is displayed again.
This patch adds a new "sectionSuggestionsForPublished" state variable
to store these suggestions.

Change-Id: I11974a66ecbb30879ad2c6a591fd3b939b41b609