e2e tests: Add "findAndWaitForRemoteActionApiRequest" method
This method intends to support requests to foreign wikis, which should
be tested separately from the local wiki. The method also expects the
domain of the foreign wiki, where the request is expected to be sent.

Additionally, the "CX_REMOTE_API_URL_TEMPLATE" should be exported to
the CLI, so that tests can properly work, according to the expected
configuration.

Change-Id: I3efeccbcd306c47f3f309efc6d40053588df8816