e2e tests: Add "findAndWaitForRecommendationApiRequest" method
This method intends to support requests to the RESTbase recommendation
API. The method also expects the domain of the foreign wiki, where the
request is expected to be sent. Currently, the method only supports
recommendation requests, but it can easily be expanded to support other
RESTbase endpoints, too.

Additionally, the "CX_RESTBASE_URL_TEMPLATE" should be exported to
the CLI, so that tests can properly work, according to the expected
configuration.

Change-Id: Id02f9260a6d67bcd41df5404c7ceb4258fbdb64a