SX: Use wiki domain code for page suggestions fetching
Currently, the request to the RESTBase API to fetch page suggestions,
uses the source language code as it is. However, this leads to failed
requests for wikis that their content language doesn't match their
wiki domain code (e.g. "nb" language code should map to the "no" wiki
domain code).

This patch fixes this by using the "getWikiDomainCode" method of
sitemapper to get and use the wiki domain code, instead of the content
language.

Change-Id: Id8ad38cf96cd971894ecf478e591931449237495