SX: Move "fetchTranslationsByStatus" vuex action to composable
This patch moves the "fetchTranslationsByStatus" from the vuex actions
(translator module) to the "useTranslationsFetch" composable and
exports it from there, so that it can be used inside any <script setup>
block.

Also, the translations are now cleared, before being (re)populated, for
a clearer syntax/implementation.

Tests are fixed accordingly.

Change-Id: I8b12fce7bdf084e12770c1721cbabfb558caf524