SX: Add previous route to application state
This patch adds a new vuex application state variable, named
"previousRoute" to store the name of the previous route (as defined in
the vue-router routes configuration). This removes the need of passing
the previous route as argument to the translation start methods and as
prop to the SXTranslationConfirmer SFC.

More importantly, this state variable will be used in subsequent patch
to properly get the event source for "dashboard_open" events, when
redirecting back to dashboard from other routes like translation
confirmer step, article search step, etc.

Bug: T353799
Change-Id: I572ee7cc801cdac7f0466c2db10034f2e773993e