SX Compare contents step: Fix section navigation
Inside the "Compare contents" step, the user should be able to navigate
among all sections, starting from missing sections. For this reason,
there are two arrow buttons to navigate to the next/previous section.

However, clicking on these buttons do not navigate the user to
next/previous section, but navigate to a randomly selected section.

This patch fixes this issue, so that click on "next"/"previous" arrow
will navigate the user to the next/previous section (starting from
missing and then moving to present sections).

Bug: T355796
Change-Id: I543d8c4f019b77dddeb5f5c4d1110c8fec623cea