SX: Remove @vue/compiler-sfc from package.json
According to package documentation:
https://www.npmjs.com/package/@vue/compiler-sfc

Note: as of 3.2.13+, this package is included as a dependency
of the main vue package and can be accessed as vue/compiler-sfc.
This means you no longer need to explicitly install this package
and ensure its version match that of Vue.

Based on the above, this patch removes this module from package.json.

Change-Id: Ib3a10caeb88dbf7c82dfaaf6144274acc1f24312