SX fetchSegmentedContent: Manually normalize page title for request
For some reason, requests to cxserver to fetch the segmented page
content for multi-word title, fail when the title is the URL uses
space between words. The requests work properly when underscore is
used inside the URL, as word separator. For example:

This URL works:
https://cxserver.wikimedia.org/v2/page/en/el/Identity_document
But this URL does NOT work:
https://cxserver.wikimedia.org/v2/page/en/el/Identity%20document

Although this issue should be also fixed in cxserver side, this
patch normalizes the page titles used for fetching the segmented
content, to avoid breaking the workflow, until the issue is fixed
in cxserver.

The same normalization already takes place in CX, so we don't need
to fix anything in that codebase.

Bug: T357769
Change-Id: I833314af22b425701e100007f575673294b03852