CX: Treat temporary users as anon for now
This patch replaces all occurrences of "isAnon" checks inside
our PHP codebase with "!isNamed" checks. This will not make
any difference for users visiting CX/SX at this point, as we
already block access to these applications for anon/temp users.
The patch will fix the visibility of some entrypoints, that
should only be visible to named users (for now).

We do have a plan about supporting anon users for CX/SX, by
creating temporary accounts for them (T353239), however, for
now we take a step back and treat temporary users as anonymous,
mostly for consistency and predictability.

Note: There are still some "isAnon" checks inside SX frontend
codebase, but they also do not have any impact, since access
to CX/SX is blocked when the configuration parameter named
"ContentTranslationEnableAnonSectionTranslation" is set to
false (which is the case for all production wikis). These checks
will be revisited and fixed, when we'll work on the feature
about supporting anon/temp users.

Bug: T344151
Change-Id: I871f54de58bb1018c7ba2b8559e09eddd3bd6e05