CX relevant languages entrypoint: Use content code instead of domain
Until now, when populating the array that includes the missing
relevant languages for the entrypoint, we only checked if the wiki
domain code for that language is valid, and then we added the code
to the missing languages array.

However, some "frequent/relevant" language codes may have a valid
wiki domain code, but they may not be a valid content language code,
and thus, they should not be included inside the missing languages
array. For example, for the "no" language code, the wiki domain
code is also "no", but the valid content language code is "nb".
Until now, the "no" language code would be added to the missing
languages array, while we should add the "nb" language code.

This patch fixes this issue by using the corresponding content
language code for the wiki domain code of each frequent language.

Bug: T322799
Change-Id: Iede1c738ce380720ffb9dc0a8222a7cfbaa8f05b