SX article language selector: Use the right set of languages for target
Inside both "Confirm a translation" and "Pick a section" step, there is
a language selector, where the user can select a different language pair
for this translation. When the user clicks on the source or target
language, a dialog opens for the user to select a new source/target
language. However, the language selector for the target languages do
not display all the available options for target language, only a subset
of them.

The reason for this issue is the following: the language selector dialog
that opens for the user to select a new source/target language, always
receives the available source languages as "input" (prop), even when
the user tries to change the target language.

The issue also exists in dashboard list language selectors, but it is
not visible, because coincidentally the source and target languages
are the same for these selectors. In "Confirm a translation" and
"Pick a section" steps, the source languages are a subset of the
available target languages, and this is why the issue is visible there.

This patch fixes this issue, by providing the right set of languages
as a prop variable, to the language selector dialog component.

Bug: T359027
Change-Id: Iae0c660ebc2ee36cdf2d11ada7c8731540986e0b