SX: Filter published/draft translations and favorites from suggestions
Until now, instead of removing the published and draft translations and
favorite suggestions from the dashboard suggestions, we removed them
from the suggestion seeds. Because of this, published and draft
translations still show up inside the suggestions list.

This patch fixes this issue by properly filtering the dashboard
suggestions to not include such suggestions.

The way that page suggestions are fetched has also changed slightly.
Instead of fetching the exact number of suggestions we need from the
API, we fetch more (24 actually) in once, since the recommendation
API enables us to do so. Some of the suggestions might be invalid
(e.g. may be draft/published translations) and we'll have to discard
them, so it's better to fetch more than the exact needed number, to
avoid multiple requests. Ultimately, we only keep 3 page suggestions per
request, so that we can populate exactly one page suggestions slice, but
no more than that, to avoid displaying suggestions from only one single
seed. Unfortunately, this is not possible for section translations, so
we do not apply a similar change there.

Also:
- removed unused "suggestion/sectionSuggestionsForArticleExists" and
"translator/getPublishedTranslationsForLanguagePair" getters.
- add logic to shuffle suggestions seeds before storing them to avoid
using the same seeds on every visit to the application

Bug: T347204
Change-Id: I4e04ad3abf76a4d7cbd002e9a4c120091e6267da