CX e2e tests: Simplify dashboard suggestion tests to be more robust
This patch updates the "dismiss page suggestion" and "refresh"
suggestions e2e tests, to only check if the corresponding suggestions
have been removed from the active suggestion list, and if the suggestion
lists are complete after dismiss/refresh.

This patch also updates the "UnifiedDashboard.page::open" method to
check if the unified dashboard page is the current page, before
redirecting, to avoid unnecessary page reloads that make the tests
slower.

Bug: T296719
Change-Id: I7e90288a4b1c4a5e391f20bb436cb04de89888d4