SX: Move "fetchLanguageTitles" from vuex action to composable
Currently, we use two vuex actions for this functionality, which is only
used inside SXTranslationConfirmer SFC. This patch replaces these vuex
actions with a composable. The composable is only used inside
SXTranslationConfirmer SFC right now, but it will be used in other
places, also, and this is why it is located inside the "composables"
directory and not inside "components/SXTranslationConfirmer" directory.

Bug: T360019
Change-Id: If1d5442e443a9d72e75cdb295a45841a5fae788f