SX: Remove useless "sx" URL parameter handling
This is a leftover from the very first days of Section Translation
development. It was used to differentiate SX URLs from CX URLs. This
has no use today, and we have no design decision for this to exist.

This patch removes it, so that the migration to Unified Dashboard is
seamless.

Change-Id: Idb5992bfe4433f34baeeac42949d4f9e9a4c3d00