SX: Create useURLHandler composable to replace urlHandler
This patch adds the "useURLHandler" composable that provides methods for
updating/retrieving URL parameters. It also stores the values of the
"page", "from", "to", "section" and "draft" URL parameters to Vue refs,
and makes sure their value always reflect the actual URL parameter. This
is done, to enable reactive access to these crucial for the application
URL parameters in a reactive way.

The following methods are returned by the composable:
- setTranslationURLParams, moved from urlHandler and adapted to store
the URL params in state (refs).
- setUrlParam, moved from urlHandler.
- getUrlParam, used in several places where the retrieval of some URL
parameter is needed, in a non-reactive way.
- clearURLParameters, used as a replacement for "replaceUrl(null)"
calls, to also clear the URL state.
- clearSectionURLParameter, moved from "useSectionSelectorClickHandler"
composable, "clearPreFilledSection" method.
- setLanguageURLParams, only used inside "useLanguageHelper".
- initializeURLState, used inside root SFC, to initalize the refs that
store the URL parameter values.

Bug: T360019
Change-Id: I223408c7624f2fcc68e3eeb31f9a1f4aace0989c