SX: Use shared composable for useDebouncedSave
When calling the debounced save translation method, we need to always
call the same method instance, otherwise several implications will
occur. Until now, we handled this by using a variable that is initially
undefined ("debouncedSaveTranslation") and is populated upon the first
call of the composable. However, this approach leads to the issue that
is described thoroughly in the linked task.

To fix this issue, this patch uses a utility function that makes the
composables "shared", that means it returns the same instance of the
composable method, until no subscribers for this composable exist.

Link to the composable method: https://github.com/vuejs/rfcs/blob/master/active-rfcs/0041-reactivity-effect-scope.md#example-a-shared-composable

In practice, we use the same instance of the composable method within
the "Pick a sentence" step, but the effect scope of that composable
is disposed when the user moves to the SXEditor. That means we cannot
use the same instance inside "Edit translation" step. This is ok, as
the pending requests from "Pick a sentence" step are properly debounced,
and are sent even if the user moves to "Edit translation" or "Publish
translation" step when an sxsave request is pending. For the SXEditor
we can use the regular "useTranslationSave" composable without debounce
and it won't make any difference: the waiting times between screen
switching is enough for the sxsave request to be sent.

Bug: T360540
Change-Id: I8b60cd17edafb1588dedfddeae3fe16d426a896d