SX SectionSuggestion: Add isListable property and use it where needed
This patch adds a new "isListable" property to the SectionSuggestion
model. This property denotes whether we should display the section
suggestion inside the dashboard suggestion list.

This property will enable us to store all the section translations
that have already been fetched inside the state, and still display
only the desired ones. This will be helpful to support the confirmation
step for draft translations of non-lead sections (where we'll need
to fetch the section translations for that page, but not display them
inside the dashboard suggestion list).

Bug: T345332
Change-Id: Ic42dca278c7789ccab5fa39a1901c0946f784570