SX: Add useLanguageTitleGroup composable
Currently, "targetPageExists" computed variable is duplicated inside two
composables, and both use the current section suggestion to compute this
boolean value.

Until now, this works well but this happens because we always make sure
that "currentSectionSuggestion" state application variable is populated
inside SXTranslationConfirmer, both for section suggestions and for page
suggestions.

However, this is counter-intuitive and confusing, the current section
suggestion should only be populated for section suggestions.

This patch uses the language title groups to define whether the target
page exists. This information is expected to be present for any kind of
translation: page suggestions, section suggestions and also draft
translations.

Finally, this patch also replaces the usage of
"getCurrentLanguageTitleGroup" getter, to use the URL parameters as
source of truth, instead of the current section suggestion Vuex state
variable.

Bug: T345332
Change-Id: I404ed733275f67e78d9290b214d5f24fcdf6cdbc