SX: Remove usage of "currentSectionSuggestion" state from two steps
This patch replaces the usage of "currentSectionSuggestion" Vuex state
variable with the "useCurrentSectionSuggestion" composable inside
"Pick a section" and "Compare the contents" steps. This is part of
our broader effort to entirely remove the "currentSectionSuggestion"
state variable and only rely on computed properties to fetch the
current section suggestion, if any.

In fact, after this patch the "currentSectionSuggestion" variable will
only be used inside the "getCurrentPage" and "getCurrentTargetPage"
getters, and thus we can easily remove it in a following patch.

Change-Id: I3179862485298b7a65780c7b95e508c26ddfe7b3