CX: Set the "publishNamespace" property when namespace is changed
Few months ago, we introduced some changes in this patch
I2515a03f5bedfda19cb61298af479f9239f0f134, to properly support section,
translation to sandbox for desktop devices. In that patch we introduced
the "publishNamespace" property inside "CXTarget" class, to hold the
value of the currently selected namespace for publishing. Initially,
this property was properly populated, but for some reason we removed
the line that populated it, in patchset 5 of the above patch. As a
result, the sandbox publishing doesn't work for section translations
on desktop.

This patch restores the setting of "publishNamespace" property, to
fix the publishing to sandbox.

Bug: T351318
Change-Id: Ie8bc44b246892f1f3a4039ae00ab07bd6ce2e0ac