SX publishing on desktop: Use returned target title for success message
Currently, the message after successful publishing on desktop uses the
target title of the current mw.cx.dm.Translation model instance to
display a message like this: "Page published at 'Target title'". This
works well with article translations for both main namespace and sandbox
publishing. When the target namespace changes to sandbox, the target
title adapts immediately to something like "User:TestUser/Title" for
article suggestions. However, the target title doesn't change when the
translation is a section translation, for sandbox publishing.

Because of that, when user publishes a section translation to sandbox,
the success message doesn't display the actual target title of the
sandbox page, but rather the target title in the main namespace.

This patch uses the target title returned by "cxpublishsection" API
to display the right target title inside the success message.

Bug: T351318
Change-Id: I15b76f0cfb28b8c802f909faa937a729c476abb0